requirements.txt
pyspark==3.3.2
fastapi==0.95.0
uvicorn==0.22.0
requests==2.26.0
sqlalchemy==1.4.40
jinja2==3.1.2

# 인증 관련 추가 라이브러리
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5

# 테스트 관련 추가 라이브러리
pytest==7.4.2
httpx==0.24.0